Move module into utils.photometry and fix minor issues.
