+++ docs/galaxies.rst
@@ -91,18 +91,6 @@ The following models are found in the `skypy.galaxies.stellar_mass` package.
   schechter_smf_mass
 
 
-Magnitude error
----------------
-
-The following models are found in the `skypy.galaxies.magnitude_error` package.
-
-.. currentmodule:: skypy.galaxies.magnitude_error
-.. autosummary::
-   :nosignatures:
-
-  rykoff_error
-
-
 Reference/API
 =============
 
@@ -115,4 +103,3 @@ Reference/API
 .. automodapi:: skypy.galaxies.spectrum
    :include-all-objects:
 .. automodapi:: skypy.galaxies.stellar_mass
-.. automodapi:: skypy.galaxies.magnitude_error
